Move blocked-tool cards under components/actions/blocked
The cards are conversation-agnostic, so they do not belong under components/assistant/conversation.
