+++ front/components/actions/blocked/PersonalAuthenticationCard.tsx
